SONARJAVA-1754 Remove deprecated 'addIssue' methods (#907)
* SONARJAVA-1754 Remove deprecated 'addIssue' methods

* SONARJAVA-1754 Standardize cost of issue as Integer for reporting

+++ java-checks/src/main/java/org/sonar/java/checks/OneClassInterfacePerFileCheck.java
@@ -41,7 +41,7 @@ public class OneClassInterfacePerFileCheck extends IssuableSubscriptionVisitor {
     int types = ((CompilationUnitTree) tree).types().size();
     if (types > 1) {
       context.addIssue(
-        -1, OneClassInterfacePerFileCheck.this, "There are " + types + " top-level types in this file; move all but one of them to other files.", (double) (types - 1));
+        -1, OneClassInterfacePerFileCheck.this, "There are " + types + " top-level types in this file; move all but one of them to other files.", types - 1);
     }
   }
