+++ java-frontend/src/test/java/org/sonar/java/xml/maven/PomCheckContextImplTest.java
@@ -111,14 +111,14 @@ public class PomCheckContextImplTest {
         reportedMessage = "onLine:" + (String) invocation.getArguments()[3];
         return null;
       }
-    }).when(sonarComponents).addIssue(any(File.class), eq(CHECK), eq(LINE), anyString(), eq((Double) null));
+    }).when(sonarComponents).addIssue(any(File.class), eq(CHECK), eq(LINE), anyString(), eq((Integer) null));
     doAnswer(new Answer<Void>() {
       @Override
       public Void answer(InvocationOnMock invocation) throws Throwable {
         reportedMessage = "onFile:" + (String) invocation.getArguments()[3];
         return null;
       }
-    }).when(sonarComponents).addIssue(any(File.class), eq(CHECK), eq(-1), anyString(), eq((Double) null));
+    }).when(sonarComponents).addIssue(any(File.class), eq(CHECK), eq(-1), anyString(), eq((Integer) null));
     doAnswer(new Answer<Void>() {
       @Override
       public Void answer(InvocationOnMock invocation) throws Throwable {
